pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "src"
version = "0.0.1"
description = "Two-tower recommendation model with transformer"
authors = [
  { name = "Stepanov Simen" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "mkdocs",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "tqdm",
    "typer",
    "torch",
    "polars",
    "grocery @ file:///D:/my_project/transformer_recsys/src/grocery",
]
requires-python = "~=3.11.0"
packages = [
    { include = "src", from = "." }, 
]

[tool.poetry.dependencies]
grocery = {develop = true}

[tool.ruff]
select = ["I", "E", "F", "W", "B", "N"]
line-length = 88 
exclude = ["data", "notebooks", ".venv", ".git", "docs"]
src = ["src"]
ignore = [ "B905", "E501", "N806"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["src"]


[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)"
]
